v3
The Network Scanner and OS Detector is a Python script that allows you to perform network scanning and detect the operating system of the hosts within a given IP range. It utilizes the Zmap and Nmap tools to accomplish these tasks. The script generates a list of IP addresses within the specified range, performs network scanning using Zmap, and then detects the operating system of each host using Nmap. The results are stored in a JSON file for further analysis.